+++ warp/context.py
@@ -3200,6 +3200,8 @@ def assert_initialized():
 
 # global entry points
 def is_cpu_available():
+    assert_initialized()
+
     return runtime.llvm